% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractVegComp.R
\name{extractVegComp_byFID}
\alias{extractVegComp_byFID}
\title{Load the attributes from vegcomp rank 1 layer from bcgw warehouse.}
\usage{
extractVegComp_byFID(bcgwUserName, bcgwPassword, FID, extractAttributes)
}
\arguments{
\item{bcgwUserName}{character, Specifies a valid user name for bcgw database.}

\item{bcgwPassword}{character, Specifies the password to the user name.}

\item{FID}{character, Specifies the feature_id in vegcomp rank1 layer.}

\item{extractAttributes}{character, The attributes that are intended to extract from vegcomp layer.
The entire list of attributes can be found at \href{https://www2.gov.bc.ca/assets/gov/farming-natural-resources-and-industry/forestry/stewardship/forest-analysis-inventory/data-management/standards/vegcomp_toc_data_dictionaryv5_2019.pdf}{here}}
}
\value{
a data table with the attributes
}
\description{
This function is to extract the attributes from vegcomp rank 1
layer from bcgw warehouse for given feature id
}
\note{
the maximum number of feature id is 1000
}
\author{
Yong Luo
}
