% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadISMC_all.R
\name{loadISMC_all}
\alias{loadISMC_all}
\title{Load all key data from ISMC Oracle database}
\usage{
loadISMC_all(
  userName,
  passWord,
  env,
  savePath = ".",
  saveFormat = "rds",
  overWrite = FALSE
)
}
\arguments{
\item{userName}{character, Specifies a valid user name in ISMC Oracle database.}

\item{passWord}{character, Specifies the password to the user name.}

\item{env}{character, Specifies which environment the data reside. Currently,
the function supports \code{INT} (intergration)
and \code {TEST} (test) environment.}

\item{savePath}{character, Specifies the path to save your outputs. If missing, the current working
directory will be choosed.}

\item{saveFormat}{character, Specifies the format for the output data.
It accepts \code{xlsx}, \code{csv}, \code{rds} and \code{txt}. Default is \code{rds}.}

\item{overWrite}{logical, Determine if the file with same name as user specifies
will be overwritten. Default is \code{FALSE}.}
}
\value{
No value returned. There are 19 tables will be saved with \code{ISMC_YYYYMMDDHHHH(ALL)_} as a prefix.
These tables are SampleSites, AccessNotes, Plots,
SampleSiteVisits, GroundSampleCrewActivities, PlotDetails, SampleMeasurements,
SmallLiveTreeTallies, TreeMeasurements, Trees, TreeDetails,
TreeDamageOccurrences, TreeLossIndicators, TreeLogAssessments,
StumpTallies, SiteNavigation,
IntegratedPlotCenter, ReferencePoint and TiePoint.
}
\description{
This function is to load all the key ground sample data from ISMC database.
}
\author{
Yong Luo
}
