% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadISMC_bySiteName.R
\name{loadISMC_bySiteName}
\alias{loadISMC_bySiteName}
\title{Load the data from ISMC Oracle database using site name(s)}
\usage{
loadISMC_bySiteName(
  userName,
  passWord,
  env,
  siteName,
  savePath = ".",
  saveName,
  saveFormat = "rdata",
  overWrite = FALSE
)
}
\arguments{
\item{userName}{character, Specifies a valid user name in ISMC Oracle database.}

\item{passWord}{character, Specifies the password to the user name.}

\item{env}{character, Specifies which environment the data reside. Currently,
the function supports \code{INT} (intergration)
and \code {TEST} (test) environment.}

\item{siteName}{character, Site name.}

\item{savePath}{character, Specifies the path to save your outputs. If missing, the current working
directory will be choosed.}

\item{saveName}{character, Specifies the save name.}

\item{saveFormat}{character, Specifies the format for the output data.
It accepts \code{rds}, \code{csv}, \code{txt}, \code{xlsx} and \code{rdata}.
Default is \code{rdata}.}

\item{overWrite}{logical, Determine if the file with same name as user specifies
will be overwritten. Default is \code{FALSE}.}
}
\value{
No value returned. There are 16 tables will be saved with prefix of \code{saveName}.
These tables are SampleSites, AccessNotes,
SampleSiteVisits, GroundSampleCrewActivities, PlotDetails, SampleMeasurements,
SmallLiveTreeTallies, TreeMeasurements, Trees, TreeDetails,
TreeDamageOccurrences, TreeLossIndicators, TreeLogAssessments,
StumpTallies, SiteNavigation,
IntegratedPlotCenter, ReferencePoint and TiePoint.
Note that if the save format is \code{xlsx}, all the tables will be saved into one workbook.
}
\description{
This function is to load the ground sample data from ISMC database using
site name(s).
}
\author{
Yong Luo
}
