% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadISMC_sizedata.R
\name{loadISMC_sizedata}
\alias{loadISMC_sizedata}
\title{Load all DBH-Height data from ISMC Oracle database}
\usage{
loadISMC_sizedata(userName, passWord, env)
}
\arguments{
\item{userName}{character, Specifies a valid user name in ISMC Oracle database.}

\item{passWord}{character, Specifies the password to the user name.}

\item{env}{character, Specifies which environment the data reside. Currently,
the function supports \code{INT} (intergration)
and \code {TEST} (test) environment.}
}
\value{
DBH and height table.
}
\description{
This function is to load all the DBHG-height data regardless of
sample site visit type. Note that this function is specifically used for
loading DBH and Height for deriving DBH-height function that used
in the PSP and VRI compiler. Hence, the broken top trees are excluded
in the dataset. The data is limited for the five major repeatedly
measured projects, i.e., 'PSP', 'M', 'Y', 'L', 'F'. The I and S samples
in the PSP projects also excluded.
}
\author{
Yong Luo
}
