% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVGIS.R
\name{loadVGIS}
\alias{loadVGIS}
\title{Load the data from VGIS Oracle database}
\usage{
loadVGIS(userName, passWord, saveThem = FALSE, savePath = ".")
}
\arguments{
\item{userName}{character, Specifies a valid user name in VGIS Oracle database.}

\item{passWord}{character, Specifies the password to the user name.}

\item{saveThem}{logical, Specifies whether the loaded data should be saved or returned.
The default value is FALSE, which means the function will not save files
for you.}

\item{savePath}{character, Specifies the path that directs to the VRI original data soruce, i.e.,
\code{//Mayhem/GIS_TIB/RDW/RDW_Data2/Work_Areas/VRI_ASCII_PROD/vri_sa}.}
}
\value{
no files
}
\description{
This function is to read the data from VGIS Oracle database
}
\author{
Yong Luo
}
