% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVGISSampleAccess.R
\name{loadVGISSampleAccess}
\alias{loadVGISSampleAccess}
\title{Load the sample access data from Oracle database}
\usage{
loadVGISSampleAccess(userName, password, saveThem = FALSE,
  savePath = file.path("."))
}
\arguments{
\item{userName}{character, Specifies a valid user name in VGIS Oracle database.}

\item{password}{character, Specifies the password to the user name.}

\item{saveThem}{logical, Specifies whether the loaded data should be saved or returned.
The default value is FALSE, which means the function will not save files
for you.}

\item{savePath}{character, Specifies the path to save your outputs, you do not need to
specify if \code{saveThem} is turned off. If missing, the current working
directory will be choosed.}
}
\value{
Depends on \code{saveThem}, if it hs turned off (i.e., \code{FALSE}). The
function returns three datasets:
\itemize{
\item{\code{} contains the crews for conducting QA;}
\item{\code{}}
}
}
\description{
This function is to load the sample header data from VGIS Oracle database. And
associated them with cluster.
}
\author{
Yong Luo
}
