% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projDate_rank1Layer.R
\name{projDate_rank1Layer}
\alias{projDate_rank1Layer}
\title{Get the project date for the rank 1 layer.}
\usage{
projDate_rank1Layer(bcgwUserName, bcgwPassword)
}
\arguments{
\item{bcgwUserName}{character, Specifies a valid user name for bcgw database.}

\item{bcgwPassword}{character, Specifies the password to the user name.}
}
\value{
project date for the rank 1 layer
}
\description{
This function is to query the project date for the rank 1 layer.
Use this as version indication.
}
\author{
Yong Luo
}
