% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleValidStat_ISMC.R
\name{sampleValidStat_ISMC}
\alias{sampleValidStat_ISMC}
\title{Obtain sample validation status by measurement year, sample type and sample project}
\usage{
sampleValidStat_ISMC(
  userName,
  passWord,
  env,
  sampleYear = NA,
  sampleType = NA,
  sampleProject = NA
)
}
\arguments{
\item{userName}{character, Specifies a valid user name in ISMC Oracle database.}

\item{passWord}{character, Specifies the password to the user name.}

\item{env}{character, Specifies which environment the data reside. Currently,
the function supports \code{INT} (intergration)
and \code {TEST} (test) environment.}

\item{sampleYear}{integer, Sample measurement year. If missing, the function takes
all the samples regardless of measurement year.}

\item{sampleType}{character, Sample type. If missing, the function takes
all the samples regardless of sample type.}

\item{sampleProject}{character, Sample project. If missing, the function takes
all the samples regardless of sample project.}
}
\value{
a data table
}
\description{
This function is to load all the table names from ISMC database.
}
\seealso{
\code{\link{tablesInISMC}}
}
\author{
Yong Luo
}
