% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablesInGYS.R
\name{tablesInGYS}
\alias{tablesInGYS}
\title{Obtain all the table names (and column names) in GYS}
\usage{
tablesInGYS(userName, passWord, columnNames = TRUE)
}
\arguments{
\item{userName}{character, Specifies a valid user name in GYS Oracle database.}

\item{passWord}{character, Specifies the password to the user name.}

\item{columnNames}{logic, Obtain column names in each table. Default is TRUE.}
}
\value{
a data table
}
\description{
This function is to load all the table names from GYS database.
}
\seealso{
\code{\link{tablesInISMC}} and \code{\link{tablesInVGIS}}
}
\author{
Yong Luo
}
