% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablesInVGIS.R
\name{tablesInVGIS}
\alias{tablesInVGIS}
\title{Obtain all the table names (and column names) in VGIS}
\usage{
tablesInVGIS(userName, passWord, columnNames = TRUE)
}
\arguments{
\item{userName}{character, Specifies a valid user name in VGIS Oracle database.}

\item{passWord}{character, Specifies the password to the user name.}

\item{columnNames}{logic, Obtain column names in each table. Default is TRUE.}
}
\value{
a data table
}
\description{
This function is to load all the table names from VGIS database.
}
\seealso{
\code{\link{tablesInISMC}} and \code{\link{tablesInGYS}}
}
\author{
Yong Luo
}
