% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadISMC_all.R
\name{loadISMC_all}
\alias{loadISMC_all}
\title{Load the data from ISMC Oracle database using sample type}
\usage{
loadISMC_all(userName, passWord, savePath = ".", saveFormat = "rds",
  overWrite = FALSE)
}
\arguments{
\item{userName}{character, Specifies a valid user name in ISMC Oracle database.}

\item{passWord}{character, Specifies the password to the user name.}

\item{savePath}{character, Specifies the path to save your outputs, you do not need to
specify if \code{saveThem} is turned off. If missing, the current working
directory will be choosed.}

\item{saveFormat}{character, Specifies the format for the output data.
It accepts \code{xlsx}, \code{csv} and \code{rds}. Default is \code{rds}.}

\item{overWrite}{logical, Determine if the file with same name as user specifies
will be overwritten. Default is \code{FALSE}.}
}
\value{
no value returned. There are 17 tables will be saved based on user's specifications.
They are SampleSites, AccessNotes, PointLocation, PlotDetails,
SampleSiteVisits, SampleMeasurements,
GroundSampleCrewActivities, SiteNavigation,
IntegratedPlotCenter, ReferencePoint, TiePoint,
SmallLiveTreeTallies, StumpTallies,
TreeMeasurements, TreeLogAssessments,
TreeDamageOccurrences and TreeLossIndicators.
}
\description{
This function is to load the ground sample data from ISMC database using
sample type.
}
\author{
Yong Luo
}
