% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadISMC_bySampleType.R
\name{loadISMC_bySampleType}
\alias{loadISMC_bySampleType}
\title{Load the data from ISMC Oracle database using sample type(s)}
\usage{
loadISMC_bySampleType(userName, passWord, sampleType, savePath = ".",
  saveFormat = "rdata", overWrite = FALSE)
}
\arguments{
\item{userName}{character, Specifies a valid user name in ISMC Oracle database.}

\item{passWord}{character, Specifies the password to the user name.}

\item{sampleType}{character, Site type or types for BC ground sample inventory, such as \code{L}, \code{M} and \code{Y}.}

\item{savePath}{character, Specifies the path to save your outputs. If missing, the current working
directory will be choosed.}

\item{saveFormat}{character, Specifies the format for the output data.
It accepts \code{xlsx} and \code{rdata}. Default is \code{rdata}.}

\item{overWrite}{logical, Determine if the file with same name as user specifies
will be overwritten. Default is \code{FALSE}.}
}
\value{
No value returned. There are 16 tables will be saved with prefix of \code{ISMC_YYYYMMDDHHHH(sampletype)_}.
These tables are SampleSites, AccessNotes,
SampleSiteVisits, GroundSampleCrewActivities, PlotDetails, SampleMeasurements,
SmallLiveTreeTallies, TreeMeasurements, Trees, TreeDetails,
TreeDamageOccurrences, TreeLossIndicators, TreeLogAssessments,
StumpTallies, SiteNavigation,
IntegratedPlotCenter, ReferencePoint and TiePoint.
}
\description{
This function is to load the ground sample data from ISMC database using
sample type(s).
}
\author{
Yong Luo
}
