% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadISMC_bySiteName.R
\name{loadISMC_bySiteName}
\alias{loadISMC_bySiteName}
\title{Load the data from ISMC Oracle database using site names}
\usage{
loadISMC_bySiteName(userName, passWord, siteName, savePath = ".",
  saveName, saveFormat = "rdata", overWrite = FALSE)
}
\arguments{
\item{userName}{character, Specifies a valid user name in ISMC Oracle database.}

\item{passWord}{character, Specifies the password to the user name.}

\item{siteName}{character, Site name.}

\item{savePath}{character, Specifies the path to save your outputs, you do not need to
specify if \code{saveThem} is turned off. If missing, the current working
directory will be choosed.}

\item{saveName}{character, Specifies the save name.}

\item{saveFormat}{character, Specifies the format for the output data.
It accepts \code{xlsx} and \code{rdata}. Default is \code{rdata}.}

\item{overWrite}{logical, Determine if the file with same name as user specifies
will be overwritten. Default is \code{FALSE}.}
}
\value{
no value returned. There are 17 tables will be saved based on user's specifications.
They are SampleSites, AccessNotes, PointLocation, PlotDetails,
SampleSiteVisits, SampleMeasurements,
GroundSampleCrewActivities, SiteNavigation,
IntegratedPlotCenter, ReferencePoint, TiePoint,
SmallLiveTreeTallies, StumpTallies,
TreeMeasurements, TreeLogAssessments,
TreeDamageOccurrences and TreeLossIndicators.
}
\description{
This function is to load the ground sample data from ISMC database using
site names.
}
\author{
Yong Luo
}
